+++ src/components/AsyncLookup/index.js
@@ -1,0 +1,4 @@
+import AsyncLookup from './AsyncLookup';
+
+export default AsyncLookup;
+
