+++ src/components/Loading/index.js
@@ -1,0 +1,4 @@
+import Loading from './Loading';
+
+export default Loading;
+
