+++ src/components/Lookup/index.js
@@ -1,0 +1,3 @@
+import Lookup from './Lookup';
+
+export default Lookup;
